requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
mysqlclient
channels==4.1.0
channels-redis==4.2.0
django-channels-jwt-auth-middleware==1.0.0
openai==1.30.1
python-dotenv==1.0.1
Pillow
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-bootstrap4==24.2
whitenoise==6.5.0
gunicorn==21.2.0
redis==5.0.4
celery
django-celery-beat
django-filter==24.2
markdown==3.6
requests==2.31.0
